Change OFX file date source from DTASOF to DTEND
I've found that, for institutions I deal with, the value of the DTASOF
field in the LEDGERBAL section is the date/time the report was
generated -- that is, the date/time the information was retrieved
over the Internet.  That's not useful to use in the filename of the
resultant archived file.

Instead, the DTEND value in the BANKTRANLIST section is the end of the
period that was queried, the "closing" date.  That is the date I want
to appear in file names.  Therefore, I here modify beancount's example
OFX importer to use this value instead of DTASOF.

I asked in the Google Beancount group whether I should create a pull
request with my changes, but the Beancount owner said not to, that
importers were to remain unsupported.  The exchange is at:
https://groups.google.com/g/beancount/c/MGEfrDd0Tgc/m/8L18I6hsBQAJ
"ofx max_date correct?"

So my intention now is to maintain my own fork of beancount at github,
accumulating any further tweaks I want in the codebase.  Substantially
different importers can land in a separate repository, but in this case
I want to track any subsequent changes to the mainline OFX importer
(possible, although it remains "unsupported").